ANDROID-10204 - Refactoring to have a more generic Card composable
